==> 1.separarArchivos.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date, timedelta

### Divido el archivo de las subastas en ventanas de 3 dias

archivo = pd.read_csv("auctions.csv")

archivo['date'] = pd.to_datetime(archivo['date'])
archivo['fecha'] = archivo['date'].dt.date

archivo_1820 = archivo[archivo['fecha'] <= date(2019,4,20)]
archivo_1921 = archivo[(archivo['fecha']>=date(2019,4,19)) & (archivo['fecha']<=date(2019,4,21))]
archivo_2022 = archivo[(archivo['fecha']>=date(2019,4,20)) & (archivo['fecha']<=date(2019,4,22))]
archivo_2123 = archivo[(archivo['fecha']>=date(2019,4,21)) & (archivo['fecha']<=date(2019,4,23))]
archivo_2224 = archivo[(archivo['fecha']>=date(2019,4,22)) & (archivo['fecha']<=date(2019,4,24))]
archivo_2325 = archivo[(archivo['fecha']>=date(2019,4,23)) & (archivo['fecha']<=date(2019,4,25))]
archivo_2426 = archivo[archivo['fecha']>=date(2019,4,24)] 

archivo_1820.to_csv('auctions_1820.csv', index=False)
archivo_1921.to_csv('auctions_1921.csv', index=False)
archivo_2022.to_csv('auctions_2022.csv', index=False)
archivo_2123.to_csv('auctions_2123.csv', index=False)
archivo_2224.to_csv('auctions_2224.csv', index=False)
archivo_2325.to_csv('auctions_2325.csv', index=False)
archivo_2426.to_csv('auctions_2426.csv', index=False)

### Divido el archivo de los eventos en ventanas de 3 dias

archivo = pd.read_csv("events.csv")

archivo['date'] = pd.to_datetime(archivo['date'])
archivo['fecha'] = archivo['date'].dt.date

archivo_1820.to_csv('events_1820.csv', index=False)
archivo_1921.to_csv('events_1921.csv', index=False)
archivo_2022.to_csv('events_2022.csv', index=False)
archivo_2123.to_csv('events_2123.csv', index=False)
archivo_2224.to_csv('events_2224.csv', index=False)
archivo_2325.to_csv('events_2325.csv', index=False)
archivo_2426.to_csv('events_2426.csv', index=False)

### Divido el archivo de las instalaciones en ventanas de 3 dias

archivo = pd.read_csv('installs.csv.gzip', compression='gzip')

archivo['created'] = pd.to_datetime(archivo['created'])
archivo['fecha'] = archivo['created'].dt.date

archivo_1820.to_csv('installs_1820.csv', index=False)
archivo_1921.to_csv('installs_1921.csv', index=False)
archivo_2022.to_csv('installs_2022.csv', index=False)
archivo_2123.to_csv('installs_2123.csv', index=False)
archivo_2224.to_csv('installs_2224.csv', index=False)
archivo_2325.to_csv('installs_2325.csv', index=False)
archivo_2426.to_csv('installs_2426.csv', index=False)

### Divido el archivo de los clicks en ventanas de 3 dias

archivo = pd.read_csv('clicks.csv.gzip', compression='gzip')

archivo['created'] = pd.to_datetime(archivo['created'])
archivo['fecha'] = archivo['created'].dt.date

archivo_1820.to_csv('clicks_1820.csv', index=False)
archivo_1921.to_csv('clicks_1921.csv', index=False)
archivo_2022.to_csv('clicks_2022.csv', index=False)
archivo_2123.to_csv('clicks_2123.csv', index=False)
archivo_2224.to_csv('clicks_2224.csv', index=False)
archivo_2325.to_csv('clicks_2325.csv', index=False)
archivo_2426.to_csv('clicks_2426.csv', index=False)
