==> 2.featuresClicks.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date

clicks = pd.read_csv("clicks.csv.gzip", compression='gzip')

clicks.head()

clicks.isnull().sum()

clicks.shape

clicks['source_id'].value_counts()

clicks['wifi_connection'].value_counts()

clicks['carrier_id'].value_counts().head(10)

clicks['os_minor'].value_counts()

clicks['os_major'].value_counts()

clicks['specs_brand'].value_counts()

clicks['timeToClick'].value_counts().head(10)

clicks['ref_type'].value_counts()

clicks['advertiser_id'].value_counts()

### Creo los archivos con features de los clicks

def crear_features_clicks(clicks_entrenar,datetime_inicio):
    clicks_entrenar['created'] = pd.to_datetime(clicks_entrenar['created'])
    
    dia_en_segundos = 86400
    clicks_entrenar['segundos'] = ((clicks_entrenar['created']-datetime_inicio).dt.seconds +
                                   (clicks_entrenar['created']-datetime_inicio).dt.days*dia_en_segundos)

    
    # Cambio advertiser_id por la probabilidad de que aparezca ese anuncio
    aviso = clicks_entrenar['advertiser_id'].value_counts(normalize=True)
    clicks_entrenar['advertiser_id'] = clicks_entrenar['advertiser_id'].replace(aviso)
        
    # Cambio source_id por la probabilidad de que aparezca ese source
    source = clicks_entrenar['source_id'].value_counts(normalize=True)
    clicks_entrenar['source_id'] = clicks_entrenar['source_id'].replace(source)
    
    # Cambio carrier_id por la probabilidad de que aparezca ese carrier
    carrier = clicks_entrenar['carrier_id'].value_counts(normalize=True)
    clicks_entrenar['carrier_id'] = clicks_entrenar['carrier_id'].replace(carrier)
    
    # Cambio ref_type por la probabilidad de que aparezca ese ref_type
    tipo = clicks_entrenar['ref_type'].value_counts(normalize=True)
    clicks_entrenar['ref_type'] = clicks_entrenar['ref_type'].replace(tipo)

    # Cambio os_major por la probabilidad de que aparezca ese os_major
    os_major = clicks_entrenar['os_major'].value_counts(normalize=True)
    clicks_entrenar['os_major'] = clicks_entrenar['os_major'].replace(os_major)

    # Cambio os_minor por la probabilidad de que aparezca ese os_minor
    os_minor = clicks_entrenar['os_minor'].value_counts(normalize=True)
    clicks_entrenar['os_minor'] = clicks_entrenar['os_minor'].replace(os_minor)


    # Agrupo el data frame por los ref_hash
    clicks_entrenar = clicks_entrenar.groupby('ref_hash').agg({'ref_hash':'count','wifi_connection':'sum',
                                                               'advertiser_id':'mean','source_id':'mean',
                                                               'carrier_id':'mean','os_major':'mean',
                                                               'os_minor':'mean','ref_type':'mean',
                                                               'segundos':'mean','timeToClick':'mean'})
    
    clicks_entrenar.columns = ['c_advertiser','c_osmin','c_t_prom','c_#clicks','c_source','c_wifi',
                                 'c_type','c_timeToClick','c_osmaj','c_carrier']
    clicks_entrenar = clicks_entrenar.reset_index('ref_hash')
          
    return clicks_entrenar

from pytz.reference import UTC
clicks_entrenar = pd.read_csv('clicks_1820.csv')
datetime_inicio = datetime(2019,4,18,0, 0, 0, tzinfo=UTC)
f_clicks = crear_features_clicks(clicks_entrenar,datetime_inicio)
f_clicks.to_csv('clicks_train_1820.csv', index=False)
f_clicks.head()

clicks_entrenar = pd.read_csv('clicks_1921.csv')
datetime_inicio = datetime(2019,4,19,0, 0, 0, tzinfo=UTC)
f_clicks = crear_features_clicks(clicks_entrenar,datetime_inicio)
f_clicks.to_csv('clicks_train_1921.csv', index=False)
f_clicks.head()

clicks_entrenar = pd.read_csv('clicks_2022.csv')
datetime_inicio = datetime(2019,4,20,0, 0, 0, tzinfo=UTC)
f_clicks = crear_features_clicks(clicks_entrenar,datetime_inicio)
f_clicks.to_csv('clicks_train_2022.csv', index=False)
f_clicks.head()

clicks_entrenar = pd.read_csv('clicks_2123.csv')
datetime_inicio = datetime(2019,4,21,0, 0, 0, tzinfo=UTC)
f_clicks = crear_features_clicks(clicks_entrenar,datetime_inicio)
f_clicks.to_csv('clicks_train_2123.csv', index=False)
f_clicks.head()

clicks_entrenar = pd.read_csv('clicks_2426.csv')
datetime_inicio = datetime(2019,4,24,0, 0, 0, tzinfo=UTC)
f_clicks = crear_features_clicks(clicks_entrenar,datetime_inicio)
f_clicks.to_csv('clicks_train_2426.csv', index=False)
f_clicks.head()
