==> 2.featuresEvents.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date, timedelta

events = pd.read_csv("events.csv.gzip", compression='gzip')

events.isnull().sum()

# Elimino las columnas que tienen mas de la mitad de los datos nulos
events = events.drop(['device_os_version','device_brand','device_city',
                           'trans_id','user_agent','carrier','device_os','connection_type'], axis=1)

events['event_id'].value_counts().head(10)

events['ref_type'].value_counts()

events['application_id'].value_counts().head(10)

events['attributed'].value_counts()

events['device_countrycode'].value_counts()

events['kind'].value_counts().head(10)

events['wifi'].value_counts()

events['device_language'].value_counts().head(10)

### Creo los archivos con features de los eventos

def crear_features_events(events_entrenar,datetime_inicio):
    events_entrenar['date'] = pd.to_datetime(events_entrenar['date'])
    
    dia_en_segundos = 86400
    events_entrenar['segundos'] = ((events_entrenar['date']-datetime_inicio).dt.seconds + 
                                     (events_entrenar['date']-datetime_inicio).dt.days*dia_en_segundos)

    
    # Cambio event_id por la probabilidad de que aparezca ese evento
    evento = events_entrenar['event_id'].value_counts(normalize=True)
    events_entrenar['event_id'] = events_entrenar['event_id'].replace(evento)
    
    # Cambio application_id por la probabilidad de que aparezca esa aplicacion
    aplicacion = events_entrenar['application_id'].value_counts(normalize=True)
    events_entrenar['application_id'] = events_entrenar['application_id'].replace(aplicacion)
      
    # Cambio kind por la probabilidad de que aparezca ese kind
    kind = events_entrenar['kind'].value_counts(normalize=True)
    events_entrenar['kind'] = events_entrenar['kind'].replace(kind)
    

    # Agrupo el data frame por los ref_hash
    events_entrenar = events_entrenar.groupby('ref_hash').agg({'ref_hash':'count','attributed':'sum',
                                                               'wifi':'sum','event_id':'mean',
                                                               'application_id':'mean','kind':'mean',
                                                               'segundos':'mean'})
    
    events_entrenar.columns = ['e_kind','e_app_id','e_event_id','e_#wifi','e_#atribuidos','e_#eventos','e_t_prom']
    events_entrenar = events_entrenar.reset_index('ref_hash')
          
    return events_entrenar
    

events_entrenar = pd.read_csv('events_1820.csv')
datetime_inicio = datetime(2019,4,18)
f_events = crear_features_events(events_entrenar,datetime_inicio)
f_events.to_csv('events_train_1820.csv', index=False)
f_events.head()

events_entrenar = pd.read_csv('events_1921.csv')
datetime_inicio = datetime(2019,4,19)
f_events = crear_features_events(events_entrenar,datetime_inicio)
f_events.to_csv('events_train_1921.csv', index=False)
f_events.head()

events_entrenar = pd.read_csv('events_2022.csv')
datetime_inicio = datetime(2019,4,20)
f_events = crear_features_events(events_entrenar,datetime_inicio)
f_events.to_csv('events_train_2022.csv', index=False)
f_events.head()

events_entrenar = pd.read_csv('events_2123.csv')
datetime_inicio = datetime(2019,4,21)
f_events = crear_features_events(events_entrenar,datetime_inicio)
f_events.to_csv('events_train_2123.csv', index=False)
f_events.head()

events_entrenar = pd.read_csv('events_2426.csv')
datetime_inicio = datetime(2019,4,24)
f_events = crear_features_events(events_entrenar,datetime_inicio)
f_events.to_csv('events_train_2426.csv', index=False)
f_events.head()
