==> 2.featuresInstalls.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date

installs = pd.read_csv("installs.csv.gzip", compression='gzip')

installs.isnull().sum()

installs.shape

installs['application_id'].value_counts().head(10)

installs['ref_type'].value_counts()

installs['attributed'].value_counts()

installs['implicit'].value_counts()

installs['device_countrycode'].value_counts()

installs['device_model'].value_counts().head(10)

installs['wifi'].value_counts()

installs['device_language'].value_counts().head(10)

### Creo lo archivos de features de las instalaciones

def crear_features_installs(installs_entrenar,datetime_inicio):
    installs_entrenar['created'] = pd.to_datetime(installs_entrenar['created'])
    
    dia_en_segundos = 86400
    installs_entrenar['segundos'] = ((installs_entrenar['created']-datetime_inicio).dt.seconds + 
                                     (installs_entrenar['created']-datetime_inicio).dt.days*dia_en_segundos)

    
    # Cambio application_id por la probabilidad de que aparezca esa aplicacion
    aplicacion = installs_entrenar['application_id'].value_counts(normalize=True)
    installs_entrenar['application_id'] = installs_entrenar['application_id'].replace(aplicacion)
        
    # Cambio device_model por la probabilidad de que aparezca ese modelo
    modelo = installs_entrenar['device_model'].value_counts(normalize=True)
    installs_entrenar['device_model'] = installs_entrenar['device_model'].replace(modelo)
    
    # Cambio device_language por la probabilidad de que aparezca ese lenguaje
    lenguaje = installs_entrenar['device_language'].value_counts(normalize=True)
    installs_entrenar['device_language'] = installs_entrenar['device_language'].replace(lenguaje)
    
    # Cambio ref_type por la probabilidad de que aparezca ese ref_type
    tipo = installs_entrenar['ref_type'].value_counts(normalize=True)
    installs_entrenar['ref_type'] = installs_entrenar['ref_type'].replace(tipo)
    
    # Agrego los wifi que faltan
    installs_entrenar['wifi'] = installs_entrenar['wifi'].fillna(False)

    # Agrupo el data frame por los ref_hash
    installs_entrenar = installs_entrenar.groupby('ref_hash').agg({'ref_hash':'count','attributed':'sum',
                                                               'wifi':'sum','implicit':'sum',
                                                               'application_id':'mean','device_model':'mean',
                                                               'device_language':'mean','ref_type':'mean',
                                                               'segundos':['min','mean','max']})
    
    installs_entrenar.columns = ['i_t_min','i_t_prom','i_t_max','i_app_id','i_#installs',
                                 'i_model','i_language','i_type','i_wifi','i_implicit','i_attributed']
    installs_entrenar = installs_entrenar.reset_index('ref_hash')
          
    return installs_entrenar

installs_entrenar = pd.read_csv('installs_1820.csv')
datetime_inicio = datetime(2019,4,18)
f_installs = crear_features_installs(installs_entrenar,datetime_inicio)
f_installs.to_csv('installs_train_1820.csv', index=False)
f_installs.head()

installs_entrenar = pd.read_csv('installs_1921.csv')
datetime_inicio = datetime(2019,4,19)
f_installs = crear_features_installs(installs_entrenar,datetime_inicio)
f_installs.to_csv('installs_train_1921.csv', index=False)
f_installs.head()

installs_entrenar = pd.read_csv('installs_2022.csv')
datetime_inicio = datetime(2019,4,20)
f_installs = crear_features_installs(installs_entrenar,datetime_inicio)
f_installs.to_csv('installs_train_2022.csv', index=False)
f_installs.head()

installs_entrenar = pd.read_csv('installs_2123.csv')
datetime_inicio = datetime(2019,4,21)
f_installs = crear_features_installs(installs_entrenar,datetime_inicio)
f_installs.to_csv('installs_train_2123.csv', index=False)
f_installs.head()

installs_entrenar = pd.read_csv('installs_2426.csv')
datetime_inicio = datetime(2019,4,24)
f_installs = crear_features_installs(installs_entrenar,datetime_inicio)
f_installs.to_csv('installs_train_2426.csv', index=False)
f_installs.head()

### Creo los archivos para testear las instalaciones

def crear_test_installs(installs_validar,datetime_validacion):
    
    dia_en_segundos = 86400
    installs_validar['created'] = pd.to_datetime(installs_validar['created'])
    
    installs_validar['segundos'] = ((installs_validar['created']-datetime_validacion).dt.seconds + 
                                     (installs_validar['created']-datetime_validacion).dt.days*dia_en_segundos)
        
    installs_validar = installs_validar.groupby('ref_hash').agg({'segundos':'mean'}).reset_index('ref_hash')  
        
    return installs_validar
   

installs_validar = pd.read_csv('installs_2123.csv')
datetime_validar = datetime(2019,4,21)
installs_validar = crear_test_installs(installs_validar,datetime_validar)
installs_validar.to_csv('installs_test_1820.csv', index=False)
installs_validar.head()

installs_validar = pd.read_csv('installs_2224.csv')
datetime_validar = datetime(2019,4,22)
installs_validar = crear_test_installs(installs_validar,datetime_validar)
installs_validar.to_csv('installs_test_1921.csv', index=False)

installs_validar = pd.read_csv('installs_2325.csv')
datetime_validar = datetime(2019,4,23)
installs_validar = crear_test_installs(installs_validar,datetime_validar)
installs_validar.to_csv('installs_test_2022.csv', index=False)

installs_validar = pd.read_csv('installs_2426.csv')
datetime_validar = datetime(2019,4,24)
installs_validar = crear_test_installs(installs_validar,datetime_validar)
installs_validar.to_csv('installs_test_2123.csv', index=False)
