==> 2.featuresSubastas2.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date


def crear_features_auctions(auctions_entrenar,datetime_inicio):

    auctions_entrenar['date'] = pd.to_datetime(auctions_entrenar['date'])
    auctions_entrenar['hora'] = auctions_entrenar['date'].dt.hour
    
    dia_en_segundos = 86400
    auctions_entrenar['segundos'] = ((auctions_entrenar['date']-datetime_inicio).dt.seconds + 
                                     (auctions_entrenar['date']-datetime_inicio).dt.days*dia_en_segundos)

    # Cambio source_id por la probabilidad de que aparezca ese source_id
    source = auctions_entrenar['source_id'].value_counts(normalize=True)
    auctions_entrenar['source_id'] = auctions_entrenar['source_id'].replace(source)
    
    # Divido por rangos los horarios en los que se produjeron las subastas
    auctions_entrenar['4-12'] = (4<=auctions_entrenar['hora']) & (auctions_entrenar['hora']<12)
    auctions_entrenar['12-20'] = (12<=auctions_entrenar['hora']) & (auctions_entrenar['hora']<20)
    auctions_entrenar['20-4'] = (20<=auctions_entrenar['hora']) | (auctions_entrenar['hora']<4)
    
    # Agrupo el data frame por los device_id
    auctions_entrenar = auctions_entrenar.groupby('device_id').agg(
        {'source_id':'mean','segundos':['min','mean','max'],'device_id':'count',
         '4-12':'sum','12-20':'sum','20-4':'sum'})
    auctions_entrenar.head()
    auctions_entrenar.columns = ['s_h20-4','s_h4-12','s_t_min','s_t_prom','s_t_max',
                                 's_source_id','s_h12-20','s_#subastas']
    auctions_entrenar = auctions_entrenar.reset_index('device_id')
    
    return auctions_entrenar

auctions_entrenar = pd.read_csv('auctions_1820.csv')
datetime_inicio = datetime(2019,4,18)
f_auctions = crear_features_auctions(auctions_entrenar,datetime_inicio)
f_auctions.to_csv('auctions_train2_1820.csv', index=False)
f_auctions.head()

auctions_entrenar = pd.read_csv('auctions_1921.csv')
datetime_inicio = datetime(2019,4,19)
f_auctions = crear_features_auctions(auctions_entrenar,datetime_inicio)
f_auctions.to_csv('auctions_train2_1921.csv', index=False)
f_auctions.head()

auctions_entrenar = pd.read_csv('auctions_2022.csv')
datetime_inicio = datetime(2019,4,20)
f_auctions = crear_features_auctions(auctions_entrenar,datetime_inicio)
f_auctions.to_csv('auctions_train2_2022.csv', index=False)
f_auctions.head()

auctions_entrenar = pd.read_csv('auctions_2123.csv')
datetime_inicio = datetime(2019,4,21)
f_auctions = crear_features_auctions(auctions_entrenar,datetime_inicio)
f_auctions.to_csv('auctions_train2_2123.csv', index=False)
f_auctions.head()

auctions_entrenar = pd.read_csv('auctions_2426.csv')
datetime_inicio = datetime(2019,4,24)
f_auctions = crear_features_auctions(auctions_entrenar,datetime_inicio)
f_auctions.to_csv('auctions_train2_2426.csv', index=False)
f_auctions.head()

def crear_test_subastas(auctions_validar,datetime_validacion):
    
    dia_en_segundos = 86400
    auctions_validar['date'] = pd.to_datetime(auctions_validar['date'])
    
    auctions_validar = auctions_validar.groupby('device_id').agg({'date':'min'}).reset_index('device_id')  
    auctions_validar['segundos'] = ((auctions_validar['date']-datetime_validacion).dt.seconds + 
                                     (auctions_validar['date']-datetime_validacion).dt.days*dia_en_segundos)
    del auctions_validar['date']
        
    return auctions_validar
    

auctions_validar = pd.read_csv('auctions_2123.csv')
datetime_validar = datetime(2019,4,21)
auctions_validar = crear_test_subastas(auctions_validar,datetime_validar)
auctions_validar.to_csv('auctions_test_1820.csv', index=False)
auctions_validar.head()

auctions_validar = pd.read_csv('auctions_2224.csv')
datetime_validar = datetime(2019,4,22)
auctions_validar = crear_test_subastas(auctions_validar,datetime_validar)
auctions_validar.to_csv('auctions_test_1921.csv', index=False)

auctions_validar = pd.read_csv('auctions_2325.csv')
datetime_validar = datetime(2019,4,23)
auctions_validar = crear_test_subastas(auctions_validar,datetime_validar)
auctions_validar.to_csv('auctions_test_2022.csv', index=False)

auctions_validar = pd.read_csv('auctions_2426.csv')
datetime_validar = datetime(2019,4,24)
auctions_validar = crear_test_subastas(auctions_validar,datetime_validar)
auctions_validar.to_csv('auctions_test_2123.csv', index=False)
