==> 2.5.crear_features2.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, date

def unir_archivos_st(auctions,events,target):

    auctions.columns = ['ref_hash','s_h20-4','s_h4-12','s_t_min','s_t_prom','s_t_max',
                        's_source_id','s_h12-20','s_#subastas']
    target.columns = ['ref_hash','segundos']
    auctions['subasta'] = 1
    
    todo = pd.merge(auctions,target,on='ref_hash',how='outer')
    todo = pd.merge(todo,events,on='ref_hash',how='left')
    
    t_max = 259200
    vacios = {'s_t_min':t_max, 's_t_prom':t_max, 's_t_max':t_max,
              'e_t_prom': t_max, 'segundos':t_max}
    todo = todo.fillna(vacios).fillna(0)
    return todo

auctions = pd.read_csv('auctions_train2_1820.csv')
events = pd.read_csv('events_train2_1820.csv')
target = pd.read_csv('auctions_test_1820.csv')
todo = unir_archivos_st(auctions,events,target)
todo.to_csv('entrenar2_st_1820.csv', index=False)
todo.head()

auctions = pd.read_csv('auctions_train2_1921.csv')
events = pd.read_csv('events_train2_1921.csv')
target = pd.read_csv('auctions_test_1921.csv')
todo = unir_archivos_st(auctions,events,target)
todo.to_csv('entrenar2_st_1921.csv', index=False)
todo.head()

auctions = pd.read_csv('auctions_train2_2022.csv')
events = pd.read_csv('events_train2_2022.csv')
target = pd.read_csv('auctions_test_2022.csv')
todo = unir_archivos_st(auctions,events,target)
todo.to_csv('entrenar2_st_2022.csv', index=False)
todo.head()

auctions = pd.read_csv('auctions_train2_2123.csv')
events = pd.read_csv('events_train2_2123.csv')
target = pd.read_csv('auctions_test_2123.csv')
todo = unir_archivos_st(auctions,events,target)
todo.to_csv('entrenar2_st_2123.csv', index=False)
todo.head()

auctions = pd.read_csv('auctions_train2_2426.csv')
events = pd.read_csv('events_train2_2426.csv')
target = pd.read_csv('ids.csv')
todo = unir_archivos_st(auctions,events,target)
todo.to_csv('predecir2_st_2426.csv', index=False)
todo.head()

def unir_archivos_sc(events,clicks,installs,target):

    todo = pd.merge(events,clicks,on='ref_hash',how='outer')
    todo = pd.merge(todo,installs,on='ref_hash',how='outer')
    todo = pd.merge(todo,target,on='ref_hash',how='outer')
    
    t_max = 259200
    vacios = {'e_t_prom': t_max, 'c_t_prom':t_max, 'c_timeToClick':t_max,
              'i_t_min':t_max, 'i_t_prom':t_max, 'i_t_max':t_max, 'segundos':t_max}
    todo = todo.fillna(vacios).fillna(0)
    del todo['i_attributed']
    return todo

events = pd.read_csv('events_train_1820.csv')
clicks = pd.read_csv('clicks_train_1820.csv')
installs = pd.read_csv('installs_train_1820.csv')
target = pd.read_csv('installs_test_1820.csv')
todo = unir_archivos_sc(events,clicks,installs,target)
todo.to_csv('entrenar_sc_1820.csv', index=False)
todo.head()

events = pd.read_csv('events_train_1921.csv')
clicks = pd.read_csv('clicks_train_1921.csv')
installs = pd.read_csv('installs_train_1921.csv')
target = pd.read_csv('installs_test_1921.csv')
todo = unir_archivos_sc(events,clicks,installs,target)
todo.to_csv('entrenar_sc_1921.csv', index=False)
todo.head()

events = pd.read_csv('events_train_2022.csv')
clicks = pd.read_csv('clicks_train_2022.csv')
installs = pd.read_csv('installs_train_2022.csv')
target = pd.read_csv('installs_test_2022.csv')
todo = unir_archivos_sc(events,clicks,installs,target)
todo.to_csv('entrenar_sc_2022.csv', index=False)
todo.head()

events = pd.read_csv('events_train_2123.csv')
clicks = pd.read_csv('clicks_train_2123.csv')
installs = pd.read_csv('installs_train_2123.csv')
target = pd.read_csv('installs_test_2123.csv')
todo = unir_archivos_sc(events,clicks,installs,target)
todo.to_csv('entrenar_sc_2123.csv', index=False)
todo.head()

events = pd.read_csv('events_train_2426.csv')
clicks = pd.read_csv('clicks_train_2426.csv')
installs = pd.read_csv('installs_train_2426.csv')
target = pd.read_csv('ids.csv')
todo = unir_archivos_sc(events,clicks,installs,target)
todo.to_csv('predecir_sc_2426.csv', index=False)
todo.head()
