==> 3.parametros_gradient.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
from sklearn.ensemble import GradientBoostingRegressor

df1 = pd.read_csv('entrenar2_st_1820.csv')
df2 = pd.read_csv('entrenar2_st_1921.csv')
df3 = pd.read_csv('entrenar2_st_2022.csv')
df4 = pd.read_csv('entrenar2_st_2123.csv')
archivos_st = [df1,df2,df3,df4]

def buscar_parametros_grd_st(grd):
    suma_errores = 0
    for df in archivos_st:
        X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
        X_train, X_test, y_train, y_test = \
        train_test_split(X, y, test_size=0.2, random_state=150)

        grd.fit(X_train,y_train)
        preds = grd.predict(X_test)
        rmse = np.sqrt(mean_squared_error(y_test, preds))
        suma_errores += rmse
       
    error_promedio = suma_errores / len(archivos_st)
    print('RMSE promedio: %f' % (error_promedio))
    return (error_promedio)

GradientBoostingClassifier(loss=’deviance’, learning_rate=0.1,
                           n_estimators=100, subsample=1.0, criterion=’friedman_mse’,
                           min_samples_split=2, min_samples_leaf=1, 
                           min_weight_fraction_leaf=0.0, max_depth=3,
                           min_impurity_decrease=0.0, min_impurity_split=None,
                           init=None, random_state=None, max_features=None, 
                           verbose=0, max_leaf_nodes=None, warm_start=False,
                           presort=’auto’, validation_fraction=0.1, 
                           n_iter_no_change=None, tol=0.0001)

errores = []
for v in [5,10,15]:
    grd = GradientBoostingRegressor(n_estimators=v)
    print('n_estimators: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [5,10,12]:
    grd = GradientBoostingRegressor(n_estimators=15,max_features=v)
    print('max_features: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,5,8,10,12]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=v)
    print('max_depth: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [0.1,0.2,0.3,0.4,0.5]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=5, learning_rate=v)
    print('larning_rate: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,4,6,8]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=3, learning_rate=0.3,
                                   min_samples_split=v)
    print('min_sample_split: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [1,2,3,4]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=3, learning_rate=0.3,
                                   min_samples_split=2, min_samples_leaf=v)
    print('min_sample_leaf: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [10,20,50,100,150]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=3, learning_rate=0.3,
                                   min_samples_split=2, min_samples_leaf=2,
                                   random_state=v)
    print('random_state: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,3,5,7,9]:
    grd = GradientBoostingRegressor(n_estimators=15, max_depth=3, learning_rate=0.3,
                                    min_samples_split=2, min_samples_leaf=2,
                                    max_leaf_nodes=v)
    print('max_leaf_nodes: %f'% (v))
    error = buscar_parametros_grd_st(grd)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

grd = GradientBoostingRegressor(n_estimators=20, max_depth=5, learning_rate=0.3,
                                    min_samples_split=2, min_samples_leaf=2,
                                    max_leaf_nodes=10)
for df in archivos_st:
    X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
    grd.fit(X,y)
f = grd.feature_importances_

df5 = pd.read_csv('entrenar_sc_1820.csv')
df6 = pd.read_csv('entrenar_sc_1921.csv')
df7 = pd.read_csv('entrenar_sc_2022.csv')
df8 = pd.read_csv('entrenar_sc_2123.csv')
archivos_sc = [df5,df6,df7,df8]

grd_sc = GradientBoostingRegressor(n_estimators=20, max_depth=5, learning_rate=0.3,
                                    min_samples_split=2, min_samples_leaf=2,
                                    max_leaf_nodes=10)
suma_errores = 0
for df in archivos_sc:
    X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
    X_train, X_test, y_train, y_test = \
    train_test_split(X, y, test_size=0.2, random_state=150)

    grd_sc.fit(X_train,y_train)
    preds = grd_sc.predict(X_test)
    rmse = np.sqrt(mean_squared_error(y_test, preds))
    suma_errores += rmse
       
error_promedio = suma_errores / len(archivos_st)
print('RMSE promedio: %f' % (error_promedio))
