==> 3.parametros_st_randomForest.ipynb <==
import pandas as pd
import numpy as np

df1 = pd.read_csv('entrenar_st_1820.csv')
df2 = pd.read_csv('entrenar_st_1921.csv')
df3 = pd.read_csv('entrenar_st_2022.csv')
df4 = pd.read_csv('entrenar_st_2123.csv')
archivos = [df1,df2,df3,df4]

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split

def buscar_parametros_randomForest_st(rf_reg):
    suma_errores = 0
    for df in archivos:
        X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
        X_train, X_test, y_train, y_test = \
        train_test_split(X, y, test_size=0.2, random_state=150)

        rf_reg.fit(X_train,y_train)
        preds = rf_reg.predict(X_test)
        rmse = np.sqrt(mean_squared_error(y_test, preds))
        suma_errores += rmse
       
    error_promedio = suma_errores / len(archivos)
    print('RMSE promedio: %f' % (error_promedio))
    return (error_promedio)

rf_reg = RandomForestRegressor(n_estimators=10, max_depth=None,
                               min_samples_split=2, min_samples_leaf=1, 
                               min_weight_fraction_leaf=0.0, max_features=’auto’,
                               max_leaf_nodes=None, min_impurity_decrease=0.0,
                               min_impurity_split=None, bootstrap=True, 
                               oob_score=False, n_jobs=None,
                               random_state=None, verbose=0,)

errores = []
for v in [5,10,15]:
    rf_reg = RandomForestRegressor(n_estimators=v)
    print('Estimators: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,5,8]:
    rf_reg = RandomForestRegressor(n_estimators=10,max_features=v)
    print('Max_features: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [10,12]:
    rf_reg = RandomForestRegressor(n_estimators=10,max_features=v)
    print('Max_features: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,5,8,10,15,20]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=v, max_features=5)
    print('Max_Depth: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [13,14,15,16,17]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=v, max_features=5)
    print('Max_Depth: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [2,3,5,8]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=v)
    print('Min_samples_split: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [7,8,9,10]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=v)
    print('Min_samples_split: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [7,8,9,10]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=v)
    print('Min_samples_split: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [1,2,3,4,5]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7, min_samples_leaf=v)
    print('Min_samples_leaf: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [0.0,0.1,0.2,0.3]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7,
                                   min_weight_fraction_leaf=v)
    print('min_weight_fraction_leaf: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [0.0,0.1,0.2,0.3]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7, min_impurity_decrease=v)
    print('min_impurity_decrease: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in ['True']:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7,bootstrap=True)
    print('bootstrap: %s'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %s' %(error,v))

errores = []
for v in [10,20,30]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7, n_jobs=v)
    print(' n_jobs: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [1,3,5,7]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=5,
                                   min_samples_split=7, random_state=v)
    print('random_state: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [5,10,20,35]:
    rf_reg = RandomForestRegressor(n_estimators=v, max_depth=14, max_features=5,
                                   min_samples_split=7)
    print('n_estimators: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

errores = []
for v in [8,9,10,11,12]:
    rf_reg = RandomForestRegressor(n_estimators=10, max_depth=14, max_features=v,
                                   min_samples_split=7)
    print('max_features: %f'% (v))
    error = buscar_parametros_randomForest_st(rf_reg)
    errores.append([error,v])
error,v = min(errores,key=lambda x: x[0])
print('Error minimo %f con valor %f' %(error,v))

### Pruebo si cambiando algunos features mi error disminuye

df1 = pd.read_csv('entrenar2_st_1820.csv')
df2 = pd.read_csv('entrenar2_st_1921.csv')
df3 = pd.read_csv('entrenar2_st_2022.csv')
df4 = pd.read_csv('entrenar2_st_2123.csv')
archivos = [df1,df2,df3,df4]

rf_reg = RandomForestRegressor(n_estimators=40, max_depth=14, max_features=9,
                                   min_samples_split=7)
suma_errores = 0
for df in archivos:
    X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
    X_train, X_test, y_train, y_test = \
    train_test_split(X, y, test_size=0.2, random_state=150)

    rf_reg.fit(X_train,y_train)
    preds = rf_reg.predict(X_test)
    rmse = np.sqrt(mean_squared_error(y_test, preds))
    print('RMSE: %f' % (rmse))
    suma_errores += rmse
       
error_promedio = suma_errores / len(archivos)
print('RMSE promedio: %f' % (error_promedio))

##### El error anterior era de 83.800 y el actual de 86.700 por lo tanto empeora.
