==> 5-predecir_adaBoost_xgb.ipynb <==
import pandas as pd
import numpy as np

from sklearn.ensemble import RandomForestRegressor
import xgboost as xgb
from sklearn.ensemble import AdaBoostRegressor

df1 = pd.read_csv('entrenar2_st_1820.csv')
df2 = pd.read_csv('entrenar2_st_1921.csv')
df3 = pd.read_csv('entrenar2_st_2022.csv')
df4 = pd.read_csv('entrenar2_st_2123.csv')
archivos_st = [df1,df2,df3,df4]

df5 = pd.read_csv('entrenar_sc_1820.csv')
df6 = pd.read_csv('entrenar_sc_1921.csv')
df7 = pd.read_csv('entrenar_sc_2022.csv')
df8 = pd.read_csv('entrenar_sc_2123.csv')
archivos_sc = [df5,df6,df7,df8]

xg_reg = xgb.XGBRegressor(objective ='reg:linear', 
            colsample_bytree = 1, learning_rate = 0.5,
            max_depth = 5, alpha = 10, n_estimators = 15) 
ada_reg = AdaBoostRegressor(base_estimator=xg_reg, random_state=10, n_estimators=5,
                        learning_rate=0.2, loss='square')

predecir = pd.read_csv('predecir2_st_2426.csv')  
for df in archivos_st:
    X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
    ada_reg.fit(X,y)
preds = ada_reg.predict(predecir.drop(['ref_hash','segundos'],axis=1))
pred_st = pd.DataFrame(preds)
pred_st.columns = ['objetivo']
pred_st = pd.DataFrame({'ref_hash': predecir['ref_hash'],
                           'obj': pred_st['objetivo']})
pred_st.head()

ada_reg = AdaBoostRegressor(base_estimator=xg_reg, random_state=10, n_estimators=5,
                        learning_rate=0.2, loss='square')

predecir = pd.read_csv('predecir_sc_2426.csv')
for df in archivos_sc:
    X,y = df.drop(['ref_hash','segundos'],axis=1),df['segundos']
    ada_reg.fit(X,y)
preds = ada_reg.predict(predecir.drop(['ref_hash','segundos'],axis=1))
pred_sc = pd.DataFrame(preds)
pred_sc.columns = ['objetivo']
pred_sc = pd.DataFrame({'ref_hash': predecir['ref_hash'],
                           'obj': pred_sc['objetivo']})
pred_sc.head()

ids = pd.read_csv('ids.csv')
print(ids.shape)
ids.head()

pred_st['ref_hash'] = pred_st['ref_hash'].astype('str')
pred_sc['ref_hash'] = pred_sc['ref_hash'].astype('str')
ids['ref_hash'] = ids['ref_hash'].astype('str')

pred_st = pred_st[pred_st['ref_hash'].isin(ids['ref_hash'])]
pred_sc = pred_sc[pred_sc['ref_hash'].isin(ids['ref_hash'])]

pred_st['ref_hash'] = pred_st['ref_hash']+'_st'
print(pred_st.shape)
pred_st.head()

pred_sc['ref_hash'] = pred_sc['ref_hash']+'_sc'
print(pred_sc.shape)
pred_sc.head()

prediccion = pred_st.append(pred_sc)
prediccion = prediccion[['ref_hash','obj']]
prediccion = prediccion.sort_values('ref_hash')
print(prediccion.shape)
prediccion.head()

prediccion.to_csv('target_adaBoost_xgb.csv', index=False)
